Preliminary Implementation
First draft of codesphinx. Compiles but has runtime errors. Cannot pick up
hardware mic which is built into laptop. Need to find which hardware device is
used for the mic on Linux.

These list the hardware sound devices on Linux, but do not provide anything
overly helpful:

	arecord -l
	aplay -l

ADDED:
Added the corpus file that was used to generate the dictionary, language model,
and vocabulary. Also added a makefile and a main function for codesphinx.

TODO:
 * Find out how to pick up mic on Linux. Using Pulse Audio.
 * Train codesphinx to work with my voice (with sphinxtrain).
 * Compare matched words against known list of words for count increment.

File(s) Modified:
	.gitignore

File(s) Added:
	Makefile
	res/codesphinx.corpus
	src/main.c